+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 #### Features
   - Contents of the cache file can now be printed in JSON format (#54)
   - New functionality for blacklisting users (#223)
+  - Space for the cache file is now pre-allocated before SCIM operations start (#197)
 
 ## v2.18 (2024-11-19)
 #### Features